sessions: Restructuring the routes.
- Created the following routes:
   /oauth2/authorize
   /oauth2/token
   /oauth2/userinfo

- More refactoring.
