+++ app/features/index.ts
@@ -34,6 +34,7 @@ import "./common/timeOutSession";
 import "./ipv/idx";
 import "./ipv/info";
 import "./ipv/json";
-import "./ipv/oauth/token";
-import "./ipv/userinfo";
+import "./ipv/oauth2/token";
+import "./ipv/oauth2/authorize";
+import "./ipv/oauth2/userinfo";
 import "./passport/start";
