+++ app/features/ipv/oauth2/authorize/controller.ts
@@ -1,0 +1,36 @@
+import { PageSetup } from "../../../../interfaces/PageSetup";
+import { Request, Response, Router } from "express";
+import { pathName } from "../../../../paths";
+import { Engine } from "../../../engine";
+
+const getAuthorize = (req: Request, res: Response): void => {
+  if (
+    req.query.response_type &&
+    req.query.redirect_uri &&
+    req.query.state &&
+    req.query.client_id
+  ) {
+    req.session.oauth = {
+      response_type: req.query.response_type,
+      redirect_uri: req.query.redirect_uri,
+      state: req.query.state,
+      client_id: req.query.client_id,
+    };
+    const engine = new Engine();
+    engine.start(req, res);
+  } else {
+    res.redirect("/error");
+  }
+};
+
+@PageSetup.register
+class SetupOAuthAuthorizeController {
+  initialise(): Router {
+    const router = Router();
+    router.get(pathName.public.oauth2.AUTHORIZE, getAuthorize);
+
+    return router;
+  }
+}
+
+export { SetupOAuthAuthorizeController, getAuthorize };
